refactor: useFormManager in hooks dir
