+++ packages/app/src/pages/ecart-rep/entreprise.tsx
@@ -1,7 +1,7 @@
 import { useRouter } from "next/router";
 
 import { useUser } from "../../hooks/useUser";
-import { useFormManager } from "../../services/apiClient/form-manager";
+import { useFormManager } from "../../hooks/useFormManager";
 import type { NextPageWithLayout } from "../_app";
 import { ClientAuthenticatedOnly } from "@components/ClientAuthenticatedOnly";
 import { RepartitionEquilibreeLayout } from "@components/layouts/RepartitionEquilibreeLayout";
